feat: implement OAuth linking and unlinking backend
- Split handlers into separate files for better organization
