+++ cmd/web/middleware.go
@@ -97,7 +97,7 @@ func (app *application) authenticate(next http.Handler) http.Handler {
 func (app *application) requireAuthentication(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		if !app.isAuthenticated(r) {
-			http.Redirect(w, r, "/user/login", http.StatusUnauthorized)
+			http.Redirect(w, r, "/user/login", http.StatusSeeOther)
 			return
 		}
