+++ internal/services/oauthservice.go
@@ -1,5 +1,7 @@
 package services
 
+import "errors"
+
 type OAuthUserInfo struct {
 	Id        string
 	Username  string
@@ -7,8 +9,16 @@ type OAuthUserInfo struct {
 	AvatarUrl string
 }
 
+type TokenExchangeOptions struct {
+	Code         string
+	RedirectPath string
+}
+
 type OAuthService interface {
 	GetRedirectUrl(state string) string
-	ExchangeCodeForToken(code string) (string, error)
+	ExchangeCodeForToken(options TokenExchangeOptions) (string, error)
 	GetUserInfo(token string) (*OAuthUserInfo, error)
+	GetRedirectUrlWithCustomCallbackPath(state string, callbackPath string) string
 }
+
+var ErrCouldNotGetOAuthUserInfo = errors.New("oauthservice: could not get oauth user info")
