+++ ui/pages/usersettingspage/modals.templ
@@ -323,7 +323,7 @@ templ authenticatedActionModal(options AuthenticatedActionModalOptions) {
 						` }
 					>
 						@components.GoogleIcon("w-5 h-5")
-						<span>{ ctxi18n.TTemplate(ctx, "user_settings_page.verify_with", map[string]string {"Provider": "GitHub"}) }</span>
+						<span>{ ctxi18n.TTemplate(ctx, "user_settings_page.verify_with", map[string]string {"Provider": "Google"}) }</span>
 					</button>
 				}
 			</form>
